refactor: move source.moduleScopes to uni-builder
