+++ .changeset/green-cats-live.md
@@ -1,0 +1,7 @@
+---
+'@rsbuild/uni-builder': patch
+'@rsbuild/webpack': patch
+'@rsbuild/shared': patch
+---
+
+refactor: move source.moduleScopes to uni-builder
