+++ packages/compat/uni-builder/src/webpack/plugins/moduleScopes.ts
@@ -1,6 +1,7 @@
 import path from 'path';
+import type { RsbuildPlugin } from '@rsbuild/webpack';
 import { ensureAbsolutePath, type ChainedConfig } from '@rsbuild/shared';
-import type { RsbuildPlugin, ModuleScopes } from '../types';
+import type { ModuleScopes } from '../../types';
 
 export const isPrimitiveScope = (
   items: unknown[],
@@ -24,15 +25,14 @@ export const applyScopeChain = (
   return options(defaults) || defaults;
 };
 
-export const pluginModuleScopes = (): RsbuildPlugin => ({
+export const pluginModuleScopes = (
+  moduleScopes?: ChainedConfig<ModuleScopes>,
+): RsbuildPlugin => ({
   name: 'plugin-module-scopes',
 
   setup(api) {
     api.modifyWebpackChain(async (chain, { CHAIN_ID }) => {
-      const config = api.getNormalizedConfig();
-      const scopeOptions = config.source.moduleScopes;
-
-      if (!scopeOptions) {
+      if (!moduleScopes) {
         return;
       }
 
@@ -40,7 +40,7 @@ export const pluginModuleScopes = (): RsbuildPlugin => ({
         '../webpackPlugins/ModuleScopePlugin'
       );
 
-      const scopes = applyScopeChain([], scopeOptions);
+      const scopes = applyScopeChain([], moduleScopes);
 
       const rootPackageJson = path.resolve(
         api.context.rootPath,
