+++ packages/compat/uni-builder/src/webpack/webpackPlugins/ModuleScopePlugin.ts
